+++ src/noteUtility.ts
@@ -1,21 +1,9 @@
-export function convertPagesObjToArray(pages: Array<{page: string}>): Array<string> {
-  let arr: Array<string> = [];
+export default function validatePages(pages: Array<{page: string}>): Array<{page: string}> {
+  let arr: Array<{page: string}> = [];
 
   pages.forEach((page) => {
-    if (page.page !== '') {
-      arr.push(page.page);
-    }
+    if (page.page !== '') arr.push(page);
   });
 
   return arr;
-}
-
-export function validatePages(pages: Array<{page: string}>): Array<{page: string}> {
-    let arr: Array<{page: string}> = [];
-
-    pages.forEach((page) => {
-        if (page.page !== '') arr.push(page);
-    });
-
-    return arr;
 }