+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+# 1.3.0
 ### Added
 - Validate author list. Will not generate citation if first, last, or both are empty.
 - Validate pages. Will note generate note citation if page is empty.